+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- New `checkoutProfile` query.
+
 ## [0.22.0] - 2020-01-30
 
 ### Added